+++ app/models/blacklight_cornell_requests/patron.rb
@@ -13,10 +13,7 @@ module BlacklightCornellRequests
     def initialize(netid)
       @netid = netid
       @record = get_folio_record
-      # @barcode = get_barcode(netid)
-      # @group = patron_group
       @preferred_service_point = get_service_point
-      @record['preferred_service_point'] = @preferred_service_point
     end
 
     def get_folio_record
@@ -24,37 +21,43 @@ module BlacklightCornellRequests
       url = ENV['OKAPI_URL']
       tenant = ENV['OKAPI_TENANT']
       @token = CUL::FOLIO::Edge.authenticate(url, tenant, ENV['OKAPI_USER'], ENV['OKAPI_PW'])
-      # Rails.logger.debug("mjc12test: Got FOLIO token #{token}")
 
-      # TODO: add error handling
-      CUL::FOLIO::Edge.patron_record(url, tenant, @token[:token], @netid)[:user]
-      # Rails.logger.debug "mjc12test: patron record for #{@netid}, #{url}, #{tenant}: #{account}"
+      response = CUL::FOLIO::Edge.patron_record(url, tenant, @token[:token], @netid)
+      user_record = response && response[:user]
+      unless user_record
+        Rails.logger.warn "Requests: No FOLIO patron record found for netid #{@netid}"
+        return nil
+      end
 
-      # Rails.logger.debug("mjc12test: Got FOLIO account #{account.inspect}")
-      # render json: account
-      # account
+      user_record
+    rescue RestClient::ExceptionWithResponse => e
+      Rails.logger.error "Requests: Failed to retrieve FOLIO patron record for netid #{@netid} (#{e.response&.code})"
+      nil
+    rescue StandardError => e
+      Rails.logger.error "Requests: Unexpected error retrieving FOLIO patron record for netid #{@netid} (#{e.class}: #{e.message})"
+      nil
     end
 
     # Use the FOLIO /service-points-users API to retrieve the patron's default service point ID,
     # if any
     def get_service_point
+      return nil unless @record
       url = "#{ENV['OKAPI_URL']}/service-points-users?query=userId==#{@record['id']}"
-      Rails.logger.debug "mjc12test2: Using URL #{url}"
+
       headers = {
         'X-Okapi-Tenant' => ENV['OKAPI_TENANT'],
         'x-okapi-token' => @token[:token],
         :accept => 'application/json',
       }
 
-      begin
-        response = RestClient.get(url, headers)
-        Rails.logger.debug "mjc12test: got PSP response #{JSON.parse(response.body)}"
-        JSON.parse(response.body).dig('servicePointsUsers', 0, 'defaultServicePointId')
-      rescue RestClient::ExceptionWithResponse => e
-        Rails.logger.debug "mjc12test: error #{e.response.code}"
-        Rails.logger.debug "mjc12test: error #{e.response.body}"
-        nil
-      end
+      response = RestClient.get(url, headers)
+      JSON.parse(response.body).dig('servicePointsUsers', 0, 'defaultServicePointId')
+    rescue RestClient::ExceptionWithResponse => e
+      Rails.logger.error "Requests: Failed to retrieve default service point for user #{@netid} (#{e.response&.code})"
+      nil
+    rescue StandardError => e
+      Rails.logger.error "Requests: Unexpected error retrieving default service point for user #{@netid} (#{e.class}: #{e.message})"
+      nil
     end
 
     def barcode
@@ -62,12 +65,12 @@ module BlacklightCornellRequests
     end
 
     def group
-      @record['patronGroup']
+      @record && @record['patronGroup']
     end
 
     def display_name
-      personal_name = @record['personal']
-      if personal_name.present?
+      personal_name = @record && @record['personal']
+      if personal_name
         [personal_name['firstName'], personal_name['lastName']].join(' ').strip
       else
         ''